rename  pyannotedbconfig.py to registry.py + rename class accordingly and move the singleton to registry.py + rename filefinder.py -> file_finder.py
