--- pyannote/database/singleton.py
@@ -1,3 +1,0 @@
-from .pyannotedbconfig import PyannoteDbConfig
-
-CFG = PyannoteDbConfig()
